Add validation for image descriptions (#65)
