+++ tests/validate.js
@@ -1,72 +1,11 @@
 const parser = require('node-html-parser');
-const path = require('path');
+const fileMatcher = require('./file-matcher');
 const fs = require('fs');
 const marked = require('marked');
 const validate = require('jsonschema').validate;
 
-
 let requiredContents = require('./required-contents');
 let unsupportedContents = require('./unsupported-contents');
-
-function matchAny(text, patterns, callback = null){
-    let result = false;
-
-    for(let pattern of patterns){
-        if(text.indexOf(pattern) != -1){
-            if(callback) callback(pattern);            
-            result = true;
-        }
-    }
-    return result;
-}
-
-function matchAll(text, patterns, callback = null){
-    let result = true;
-
-    for(let pattern of patterns){
-        if(text.indexOf(pattern) == -1){   
-            if(callback) callback(pattern);         
-            result = false;
-        }
-    }
-    return result;
-}
-
-/**
- * 
- * @param {The directory from which to start a recursive search} startPath 
- * @param {The file name that should be looked for} searchPattern 
- * @param {An array of paths that should be excluded from the search} excludePatterns 
- * @param {The matching files as recursion parameter} matchingFiles 
- */
-function findAllFiles(startPath, searchPattern, excludePatterns = [], matchingFiles = []) {    
-    if(matchAny(startPath, excludePatterns)){
-        // console.log("Excluding directory " + startPath);
-        return matchingFiles;
-    }
-    
-    // console.log('Starting from dir ' + startPath + '/');
-
-    if (!fs.existsSync(startPath)) {
-        console.log("❌ Directory doesn't exist ", startPath);
-        return;
-    }
-
-    var files = fs.readdirSync(startPath);
-    for (var i = 0; i < files.length; i++) {
-        var filename = path.join(startPath, files[i]);
-        var stat = fs.lstatSync(filename);
-        if (stat.isDirectory()) {
-            findAllFiles(filename, searchPattern, excludePatterns, matchingFiles);
-        } else if (!matchAny(filename, excludePatterns) && filename.indexOf(searchPattern) >= 0) {
-            // console.log('-- found: ', filename);
-            matchingFiles.push(filename);
-        };
-    };
-
-    return matchingFiles;
-};
-
 
 const excludePatterns = [".git", "/portenta-h7/metadata.json", "/template/"];
 let errorOccurred = false;
@@ -74,7 +13,7 @@ let errorOccurred = false;
 /**
  * Verify that all meta data files are valid JSON and contain the correct attributes
  */
-let metaDataFiles = findAllFiles('../content/', 'metadata.json', excludePatterns);
+let metaDataFiles = fileMatcher.findAllFiles('../content/', 'metadata.json', excludePatterns);
 
 if(metaDataFiles.length == 0) {
     console.log("❌ No metadata files found.");
@@ -113,7 +52,7 @@ metaDataFiles.forEach(path => {
  * Verify that SVG images don't contain embedded images
  */
 
- let svgFiles = findAllFiles('../content/', '.svg');
+ let svgFiles = fileMatcher.findAllFiles('../content/', '.svg');
  svgFiles.forEach(path => {     
     const rawData = fs.readFileSync(path);
     if(rawData.includes("<image ")){
@@ -127,11 +66,12 @@ metaDataFiles.forEach(path => {
     }
  });
 
+
 /**
  * Verify that the content files contain the necessary data
  */
 
-let contentFiles = findAllFiles('../content/', 'content.md', excludePatterns);
+let contentFiles = fileMatcher.findAllFiles('../content/', 'content.md', excludePatterns);
 
 if(contentFiles.length == 0) {
     console.log("❌ No content files found.");
@@ -143,28 +83,33 @@ contentFiles.forEach(path => {
         let rawData = fs.readFileSync(path);
         let markdown = rawData.toString();
 
-        if(!matchAll(markdown, requiredContents, (match)=> {
+        if(!fileMatcher.matchAll(markdown, requiredContents, (match)=> {
             console.log("❌ " + path + " doesn't contain the required content : " + match);
         })){
             errorOccurred = true;
         }
 
         let htmlContent = marked(markdown);        
-        if(matchAny(htmlContent,unsupportedContents, (match) => {
+        if(fileMatcher.matchAny(htmlContent,unsupportedContents, (match) => {
             console.log("❌ " + path + " contains unsupported content : " + match);
         })){
             errorOccurred = true;
-        }        
-
-        const imageSourceRegex = /!\[([^\[]+)\](\((.*)\))/gm;
-        let imageMatch;
-        while ((imageMatch = imageSourceRegex.exec(markdown))) {
-            const imagePath = imageMatch[3];
+        }
+        
+        const htmlDoc = parser.parse(htmlContent);
+        let images = htmlDoc.querySelectorAll("img");        
+        images.forEach(image => {
+            const imagePath = image.attributes.src;
             if(imagePath.startsWith("/") || imagePath.startsWith("~")){
                 console.log("❌ Image uses an absolute path: " + imagePath + " in " + path);
                 errorOccurred = true;
-            }            
-        }
+            }             
+            const imageDescription = image.attributes.alt;
+            if(imageDescription.split(" ").length <= 1){
+                console.log("❌ Image doesn't have a description: " + imagePath + " in " + path);
+                errorOccurred = true;
+            }
+        });
         
     } catch (error) {   
         console.log(error);
